#139 - fix
